worked on Payment entity and its layers. Testing and Error handling.

+++ src/main/java/com/turf_time/dtos/PaymentDto.java
@@ -36,5 +36,5 @@ public class PaymentDto {
 	private String paymentMethod;
 
 	@NotNull(message = "Booking ID is required")
-	private Long bookingId;
+	private Integer bookingId;
 }
